fix query role with Enum problem

+++ Services/User/IW/Services/UserService.cs
@@ -41,7 +41,7 @@ namespace IW.Services
         public async Task<string> LogIn(CreateUser model)
         {
             var user = await _unitOfWork.Users.FindByCondition(u => u.Name == model.Name);
-            Role? role = !Equals(user,null)? user.Role: await _unitOfWork.Roles.GetById(ROLE.User); 
+            Role? role = !Equals(user,null)? user.Role: await _unitOfWork.Roles.GetById((int)ROLE.User); 
 
             User newUser = model.Adapt<User>();
             newUser.RoleId = role.Id;
